requirements.txt
amqp==2.5.1
ansible==2.8.6
asgiref==3.2.2
asn1crypto==0.24.0
autopep8==1.4.4
backcall==0.1.0
backports.csv==1.0.7
bcrypt==3.1.7
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.5
django-celery-results==1.1.2
django-cors-headers==3.1.1
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-filter==2.2.0
django-formtools==2.1
django-guardian==2.1.0
django-import-export==1.2.0
django-pure-pagination==0.3.0
django-reversion==3.0.4
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
et-xmlfile==1.0.1
future==0.17.1
gnureadline==8.0.0
httplib2==0.13.1
idna==2.8
importlib-metadata==0.23
ipython==7.8.0
ipython-genutils==0.2.0
itypes==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.1
kombu==4.6.5
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
msgpack==0.6.2
mysqlclient==1.4.4
odfpy==1.4.0
openpyxl==2.6.3
paramiko==2.6.0
parso==0.5.1
pbr==5.4.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pipenv==2018.11.26
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.4.2
PyMySQL==0.9.3
PyNaCl==1.3.0
pytz==2019.2
PyYAML==5.1.2
rabbitmq==0.2.0
redis==3.3.8
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
stevedore==1.31.0
tablib==0.13.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
virtualenv==16.7.4
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.7
xlrd==1.2.0
XlsxWriter==1.2.1
xlwt==1.3.0
yagmail==0.11.220
zipp==0.6.0
